requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
google-generativeai>=0.3.0

